Rename and make wrappers for remaining deprecated distributions ref #12688
