+++ modules/stochastic_tools/include/distributions/JohnsonSBDistribution.h
@@ -9,39 +9,14 @@
 
 #pragma once
 
-#include "NormalDistribution.h"
+#include "JohnsonSB.h"
 
 /**
- * A class used to generate a Johnson SB distribution
+ * A deprecated wrapper class used to generate a Johnson SB distribution
  */
-class JohnsonSBDistribution : public NormalDistribution
+class JohnsonSBDistribution : public JohnsonSB
 {
 public:
   static InputParameters validParams();
-
   JohnsonSBDistribution(const InputParameters & parameters);
-
-  virtual Real pdf(const Real & x) const override;
-  virtual Real cdf(const Real & x) const override;
-  virtual Real quantile(const Real & p) const override;
-
-  static Real
-  pdf(const Real & x, const Real & a, const Real & b, const Real & alpha_1, const Real & alpha_2);
-  static Real
-  cdf(const Real & x, const Real & a, const Real & b, const Real & alpha_1, const Real & alpha_2);
-  static Real quantile(
-      const Real & p, const Real & a, const Real & b, const Real & alpha_1, const Real & alpha_2);
-
-protected:
-  /// The lower location parameter, a
-  const Real & _lower;
-
-  /// The upper location parameter, b
-  const Real & _upper;
-
-  /// The first shape parameter, alpha_1
-  const Real & _alpha_1;
-
-  /// The second shape parameter, alpha_2
-  const Real & _alpha_2;
 };
