+++ modules/stochastic_tools/include/distributions/TruncatedNormalDistribution.h
@@ -9,42 +9,14 @@
 
 #pragma once
 
-#include "NormalDistribution.h"
+#include "TruncatedNormal.h"
 
 /**
- * A class used to generate a truncated normal distribution
+ * A deprecated wrapper class used to generate a truncated normal distribution
  */
-class TruncatedNormalDistribution : public NormalDistribution
+class TruncatedNormalDistribution : public TruncatedNormal
 {
 public:
   static InputParameters validParams();
-
   TruncatedNormalDistribution(const InputParameters & parameters);
-
-  virtual Real pdf(const Real & x) const override;
-  virtual Real cdf(const Real & x) const override;
-  virtual Real quantile(const Real & p) const override;
-
-  static Real pdf(const Real & x,
-                  const Real & mean,
-                  const Real & std_dev,
-                  const Real & lower_bound,
-                  const Real & upper_bound);
-  static Real cdf(const Real & x,
-                  const Real & mean,
-                  const Real & std_dev,
-                  const Real & lower_bound,
-                  const Real & upper_bound);
-  static Real quantile(const Real & p,
-                       const Real & mean,
-                       const Real & std_dev,
-                       const Real & lower_bound,
-                       const Real & upper_bound);
-
-protected:
-  /// The lower bound for the distribution
-  const Real & _lower_bound;
-
-  /// The upper bound for the distribution
-  const Real & _upper_bound;
 };
