+++ modules/stochastic_tools/src/distributions/UniformDistribution.C
@@ -9,74 +9,14 @@
 
 #include "UniformDistribution.h"
 
-registerMooseObjectAliased("StochasticToolsApp", UniformDistribution, "Uniform");
 registerMooseObjectReplaced("StochasticToolsApp", UniformDistribution, "07/01/2020 00:00", Uniform);
 
 InputParameters
 UniformDistribution::validParams()
 {
-  InputParameters params = Distribution::validParams();
-  params.addClassDescription("Continuous uniform distribution.");
-  params.addParam<Real>("lower_bound", 0.0, "Distribution lower bound");
-  params.addParam<Real>("upper_bound", 1.0, "Distribution upper bound");
-  return params;
+  return Uniform::validParams();
 }
 
-UniformDistribution::UniformDistribution(const InputParameters & parameters)
-  : Distribution(parameters),
-    _lower_bound(getParam<Real>("lower_bound")),
-    _upper_bound(getParam<Real>("upper_bound"))
+UniformDistribution::UniformDistribution(const InputParameters & parameters) : Uniform(parameters)
 {
-  if (_lower_bound >= _upper_bound)
-    mooseError("The lower bound is larger than the upper bound!");
-}
-
-Real
-UniformDistribution::pdf(const Real & x, const Real & lower_bound, const Real & upper_bound)
-{
-  if (x < lower_bound || x > upper_bound)
-    return 0.0;
-  else
-    return 1.0 / (upper_bound - lower_bound);
-}
-
-Real
-UniformDistribution::cdf(const Real & x, const Real & lower_bound, const Real & upper_bound)
-{
-  if (x < lower_bound)
-    return 0.0;
-  else if (x > upper_bound)
-    return 1.0;
-  else
-    return (x - lower_bound) / (upper_bound - lower_bound);
-}
-
-Real
-UniformDistribution::quantile(const Real & y, const Real & lower_bound, const Real & upper_bound)
-{
-  if (y < 0 || y > 1)
-    ::mooseError("The cdf_value provided is out of range 0 to 1.");
-  else
-    return y * (upper_bound - lower_bound) + lower_bound;
-}
-
-Real
-UniformDistribution::pdf(const Real & x) const
-{
-  TIME_SECTION(_perf_pdf);
-  return pdf(x, _lower_bound, _upper_bound);
-}
-
-Real
-UniformDistribution::cdf(const Real & x) const
-{
-  TIME_SECTION(_perf_cdf);
-  return cdf(x, _lower_bound, _upper_bound);
-}
-
-Real
-UniformDistribution::quantile(const Real & y) const
-{
-  TIME_SECTION(_perf_quantile);
-  return quantile(y, _lower_bound, _upper_bound);
 }
